osd: changed PGLOG to use recycle entries The WAL is implamnted on top of RocksDB key-data store where each change was logged with key being the eversion and the data being the changed ONode. Every update was written to the LSM tree and when level-0 was filled it was dumped to disk and a new level was open and so on and so forth. When we were done with a log entry we trimmed it - which was translated to another LSM write This forced very expenive compcations process even so we never read old log entries after they were trimmed.
This commit changed PGLog behavior to use a set of recycled keys for the all the log entries where keys are recylced after trim/remove and so the number of keys stays very short and we are now able to keep the full set in level-0 and never need to create disk-levels so no compaction is ever needed

By making a contribution to this project, I certify that:

The contribution was created in whole or in part by me and I have the right to submit it under the open source license indicated in the file; or
The contribution is based upon previous work that, to the best of my knowledge, is covered under an appropriate open source license and I have the right under that license to submit that work with modifications, whether created in whole or in part by me, under the same open source license (unless I am permitted to submit under a different license), as indicated in the file; or
The contribution was provided directly to me by some other person who certified (a), (b) or (c) and I have not modified it.
I understand and agree that this project and the contribution are public and that a record of the contribution (including all personal information I submit with it, including my sign-off) is maintained indefinitely and may be redistributed consistent with this project or the open source license(s) involved.

Signed-off-by: Gabriel BenHanokh <benhanokh@gmail.com>